requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
Django==4.2
django-cors-headers==3.14.0
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
djangorestframework==3.14.0
elastic-transport==8.4.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
